숫자 포맷팅과 인바디 페이지 제작

+++ Assets/Scripts/DataController.cs
@@ -116,7 +116,7 @@ public class DataController : MonoBehaviour
     {
         return healthMulRate;
     }
-    public float getAllHealth()
+    public int getAllHealth()
     {
         return allHealth;
     }
